requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
werkzeug>=3.0.0
